Install react icons. Move heatmapUtils.
