+++ package.json
@@ -16,6 +16,7 @@
     "daisyui": "^5.0.5",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-icons": "^5.5.0",
     "tailwindcss": "^4.0.14"
   },
   "devDependencies": {
