+++ src/components/Heatmap/heatmapUtils.ts
